refactor(forgetting): relocate ForgetPolicy/PruneStats out of traits.rs + correct importance-formula doc
`traits.rs` should declare the consumer behaviour contracts (the five
injected traits), not house concrete forgetting-layer types. `ForgetPolicy`
(a config struct with validation) and `PruneStats` (a plain output struct)
are neither traits, so move both into a new `forgetting/types.rs` alongside
their `Default`/inherent impls and their unit tests. The crate-root re-export
is preserved (`pub use forgetting::{ForgetPolicy, PruneStats}` replaces the
`pub use traits::{...}` entries), so `memory_engine::ForgetPolicy` /
`memory_engine::PruneStats` are unchanged — only the deep-path
`memory_engine::traits::ForgetPolicy` import in the MCP handler (and the test
/example/bench deep-paths) had to move to the crate root.

While moving the struct, fix its drifted doc (#380): the old comment stated an
un-normalized two-factor formula (`recency × decay + frequency × log(...) + …`)
that contradicts the actual `compute_importance` impl, which normalizes each of
the four signals to [0,1] before the weighted sum (frequency via ln(101),
graph degree via ln(51)). The doc now mirrors the impl and notes the max score
equals the sum of the four weights, cross-referencing `compute_importance` to
prevent future drift.

Behavior-preserving: only module placement, import paths, and a doc comment
changed.

Closes #351
Closes #380
Part of #255
